Change user response format

+++ db/migrate/20180530200713_create_meteors.rb
@@ -1,7 +1,7 @@
 class CreateMeteors < ActiveRecord::Migration[5.2]
   def change
     create_table :meteors do |t|
-      t.float :value
+      t.integer :value
       t.string :description
       t.belongs_to :user
